chore: improve code quality with gosec G115 overflow checks
- Re-enable gosec linter with G115 integer overflow checks
- Add nilerr and unconvert linters
- Add SafeInt32 and SafeInt32FromValue helper functions in state_helpers.go
  for safe int64 to int32 conversion with overflow checking
- Update all datasources and resources to use SafeInt32 helpers instead of
  direct int32() casts
- Remove unnecessary type conversions flagged by unconvert linter
- Add #nosec comments to test files where values are known to be safe
  (controlled test data, API-generated IDs)
- All acceptance tests pass against local Netbox container

+++ internal/resources/prefix_resource.go
@@ -277,7 +277,7 @@ func (r *PrefixResource) Update(ctx context.Context, req resource.UpdateRequest,
 	})
 
 	// Update the prefix
-	prefix, httpResp, err := r.client.IpamAPI.IpamPrefixesUpdate(ctx, int32(id)).WritablePrefixRequest(*prefixRequest).Execute()
+	prefix, httpResp, err := r.client.IpamAPI.IpamPrefixesUpdate(ctx, id).WritablePrefixRequest(*prefixRequest).Execute()
 	if err != nil {
 		resp.Diagnostics.AddError(
 			"Error updating prefix",
